将数据源更换为动漫之家 下载管理器半成品

+++ app/src/main/java/com/tongming/manga/mvp/view/activity/HistoryActivity.java
@@ -4,17 +4,19 @@ import android.content.DialogInterface;
 import android.content.Intent;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v7.app.AlertDialog;
+import android.support.v7.widget.GridLayoutManager;
+import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
 import android.view.View;
-import android.widget.AdapterView;
-import android.widget.GridView;
 
 import com.orhanobut.logger.Logger;
 import com.tongming.manga.R;
+import com.tongming.manga.cusview.SpaceItemDecoration;
 import com.tongming.manga.mvp.base.BaseActivity;
 import com.tongming.manga.mvp.bean.HistoryComic;
 import com.tongming.manga.mvp.presenter.HistoryPresenterImp;
-import com.tongming.manga.mvp.view.adapter.ComicAdapter;
+import com.tongming.manga.mvp.view.adapter.RVComicAdapter;
+import com.tongming.manga.util.CommonUtil;
 
 import java.util.List;
 
@@ -26,10 +28,14 @@ import butterknife.BindView;
 public class HistoryActivity extends BaseActivity implements IHistoryView {
     @BindView(R.id.toolbar)
     Toolbar toolbar;
-    @BindView(R.id.gv_history)
-    GridView gvHistory;
+    //    @BindView(R.id.gv_history)
+//    GridView gvHistory;
+    @BindView(R.id.rv_history)
+    RecyclerView rvHistory;
     @BindView(R.id.fab_clear)
     FloatingActionButton fabClear;
+    private List<HistoryComic> comics;
+    private RVComicAdapter adapter;
 
     @Override
     protected int getLayoutId() {
@@ -39,10 +45,6 @@ public class HistoryActivity extends BaseActivity implements IHistoryView {
     @Override
     protected void initView() {
         initToolbar(toolbar);
-        if (presenter == null) {
-            presenter = new HistoryPresenterImp(this);
-        }
-        ((HistoryPresenterImp) presenter).queryAllHistory(this);
         fabClear.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -52,18 +54,14 @@ public class HistoryActivity extends BaseActivity implements IHistoryView {
                         .setPositiveButton("确定", new DialogInterface.OnClickListener() {
                             @Override
                             public void onClick(DialogInterface dialog, int which) {
-                                ((HistoryPresenterImp) presenter).deleteAllHistory(HistoryActivity.this);
+                                if (presenter != null) {
+                                    ((HistoryPresenterImp) presenter).deleteAllHistory(HistoryActivity.this);
+                                }
                             }
                         }).setNegativeButton("取消", null).show();
             }
         });
-    }
-
-    @Override
-    public void onQuery(final List<HistoryComic> comics) {
-        Logger.d("查询历史记录");
-        gvHistory.setAdapter(new ComicAdapter(comics, this, ComicAdapter.HISTORY_COMIC));
-        gvHistory.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+        /*gvHistory.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> parent, View view, int position, long id) {
                 Intent intent = new Intent(HistoryActivity.this, ComicDetailActivity.class);
@@ -88,7 +86,68 @@ public class HistoryActivity extends BaseActivity implements IHistoryView {
                         .show();
                 return true;
             }
+        });*/
+
+    }
+
+    private void initRecyclerView() {
+        rvHistory.setLayoutManager(new GridLayoutManager(this, 3));
+        rvHistory.addItemDecoration(new SpaceItemDecoration(CommonUtil.dip2px(this, 10), true));
+        if (adapter == null) {
+            adapter = new RVComicAdapter(comics, this, RVComicAdapter.HISTORY_COMIC);
+        }
+        rvHistory.setAdapter(adapter);
+        adapter.setOnItemClickListener(new RVComicAdapter.OnItemClickListener() {
+            @Override
+            public void onItemClick(View view, int position) {
+                Intent intent = new Intent(HistoryActivity.this, ComicDetailActivity.class);
+                HistoryComic comic = comics.get(position);
+                String name = comic.getName();
+                intent.putExtra("url", comic.getUrl()).putExtra("name", name.endsWith("漫画") ? name.replace("漫画", "") : name);
+                startActivity(intent);
+            }
         });
+        adapter.setOnItemLongClickListener(new RVComicAdapter.OnItemLongClickListener() {
+            @Override
+            public boolean onItemLongClick(View view, final int position) {
+                new AlertDialog.Builder(HistoryActivity.this)
+                        .setTitle("注意")
+                        .setMessage("确定要删除" + comics.get(position).getName() + "吗?")
+                        .setPositiveButton("确定", new DialogInterface.OnClickListener() {
+                            @Override
+                            public void onClick(DialogInterface dialog, int which) {
+                                ((HistoryPresenterImp) presenter).deleteHistoryByName(HistoryActivity.this, comics.get(position).getName());
+                            }
+                        }).setNegativeButton("取消", null)
+                        .show();
+                return true;
+            }
+        });
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        queryHistory();
+    }
+
+    private void queryHistory() {
+        if (presenter == null) {
+            presenter = new HistoryPresenterImp(this);
+        }
+        ((HistoryPresenterImp) presenter).queryAllHistory(this);
+    }
+
+    @Override
+    public void onQuery(List<HistoryComic> comics) {
+        if (this.comics == null) {
+            this.comics = comics;
+            initRecyclerView();
+        } else {
+            this.comics.clear();
+            this.comics.addAll(comics);
+            adapter.notifyDataSetChanged();
+        }
     }
 
     @Override
